main/ts-srv: userspace touchscreen driver for the HP TouchPad
The touchscreen controller in the TouchPad is a Cypress CY8CTMA395
microcontroller presumably running custom firmware. This microcontroller is
connected to over UART and I2C, with the touch data being transmitted over UART.
There also appears to be an internal JTAG connection for flashing the
controller. Most of the touch calculations are done in userspace, by the driver
included in this commit.

This driver was taken from Android (specifically Evervolv), and patched to run
on without the Android NDK and to fix some compatibility issues with libinput.